Move tests to their final place
